Revert "[FSSDK-11371] parsing holdout section from the datafile (#577)"
This reverts commit 9d635dc64f669b342b26aced845f6b6c1c45deb5.
